# docs/swagger.yaml
basePath: /
definitions:
  auth.LoginRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  auth.LoginResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  auth.LogoutRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  auth.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
...
  marketplace.LinkShopRequest:
    properties:
      shop_id:
        type: string
    required:
    - shop_id
    type: object
  marketplace.LinkShopStartResponse:
    properties:
      connected:
        type: boolean
      shop_id:
        type: string
    type: object
  marketplace.LogisticChannelDTO:
    properties:
      code:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  marketplace.LogisticChannelsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/marketplace.LogisticChannelDTO'
        type: array
      message:
        type: string
    type: object
  marketplace.ShopDetailResponse:
    properties:
      data:
        properties:
          access_level:
            type: string
          country:
            type: string
          currency:
            type: string
          marketplace:
            type: string
          shop_id:
            type: string
          shop_name:
            type: string
        type: object
      message:
        type: string
    type: object
...
  response.ErrorResponse:
    properties:
      code:
        type: integer
      errors:
        description: For validation errors
        items:
          type: string
        type: array
      message:
        type: string
    type: object
  response.SuccessResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
  description: Warehouse Management System API for authentication, order processing,
    and marketplace integrations.
  title: WMS API
  version: "1.0"
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate an internal user and return access and refresh tokens.
      parameters:
      - description: Login payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/auth.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Login user
      tags:
      - Auth
  /api/auth/logout:
    post:
      consumes:
      - application/json
      description: Revoke a refresh token session.
      parameters:
      - description: Logout payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.LogoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Logout user
      tags:
      - Auth
  /api/auth/refresh:
    post:
      consumes:
      - application/json
      description: Exchange a valid refresh token for a new access token pair.
      parameters:
      - description: Refresh token payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/auth.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Refresh access token
      tags:
      - Auth
  /api/integrations/marketplace/oauth/callback:
    get:
      description: Complete marketplace OAuth callback using code, shop ID, and state
        query parameters.
      parameters:
      - description: Authorization code
        in: query
        name: code
        required: true
        type: string
      - description: Marketplace shop ID
        in: query
        name: shop_id
        required: true
        type: string
      - description: OAuth state
        in: query
        name: state
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Complete OAuth callback
      tags:
      - Marketplace
  /api/integrations/marketplace/shops/{shopID}:
    get:
      description: Retrieve marketplace shop metadata for a connected shop.
      parameters:
      - description: Marketplace shop ID
        in: path
        name: shopID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/marketplace.ShopDetailResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get connected shop detail
      tags:
      - Marketplace
  /api/integrations/marketplace/shops/{shopID}/logistic/channels:
    get:
      description: Fetch available logistic channels for a connected marketplace shop.
      parameters:
      - description: Marketplace shop ID
        in: path
        name: shopID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/marketplace.LogisticChannelsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get logistic channels
      tags:
      - Marketplace
  /api/integrations/marketplace/shops/connect/start:
    post:
      consumes:
      - application/json
      description: Start the one-step marketplace shop linking flow and persist credentials.
      parameters:
      - description: Marketplace shop payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/marketplace.LinkShopRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/marketplace.LinkShopStartResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Connect marketplace shop
      tags:
      - Marketplace
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Use the format `Bearer <access-token>`.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"